added short summary (click here for content)
For my contribution, I was responsible for the code:
Class MorraApp
In class MorraGameManager:
	MorraGameManager constructor
	playGame method
	printGamesSummary method
In class MorraGame:
	MorraGame constructor
	calcScore method
	setPlayerTeam method
	getPlayerTeam method
	toString method
During the project, I helped to fix grammatical errors that could occur in the code, and contributed to writing documents submitted by the group. I also helped to overcome issues we faced in the creation of the program, such as the fix for what to do to prevent the user from playing more games than what the array could store, and the fix for an error being thrown when the scanner was expecting a nextInt but received a String.